BQP: Fix Deliverable + fix role + fix status ProcessOption

+++ modules/backend-dossiermgt/backend-dossiermgt-api/src/main/java/org/opencps/dossiermgt/model/ProcessOptionModel.java
@@ -360,14 +360,14 @@ public interface ProcessOptionModel extends BaseModel<ProcessOption>,
 	 *
 	 * @return the status of this process option
 	 */
-	public long getStatus();
+	public int getStatus();
 
 	/**
 	 * Sets the status of this process option.
 	 *
 	 * @param status the status of this process option
 	 */
-	public void setStatus(long status);
+	public void setStatus(int status);
 
 	@Override
 	public boolean isNew();
